Create PoS General FAQ
A simplified general FAQ for non technical people who are new to Decred.
Based on the most frequently asked questions in the slack channel.

The idea and structure of this FAQ is to introduce new terms along the way and build on them as to be not too confusing.

Please read and if there's any mistakes I'll change it. As I'm certain there are details that I missed or have gotten wrong.
Also I hope the layout works as this is my first commit of this kind.