Update block header examples.
This modifies the block header specification so that the comments on the
example encoded block header line up with each other to ease readability
and also corrects the PoolSize reference.  Also, it updates the example
encoded block header as raw bytes to use the available space so it more
closely matches the example getwork response below it.