Change file size sentences and add SPV
As noted by chappjc: "the size dcrdata reports is the sum of the sizes of all serialized blocks, as opposed to the disk space occupied by the database. There is quite a bit of metadata and other DB overhead that adds to the size on disk"

I changed the final sentences to reflect this more clearly. I also added the notion of SPV, since it doesn't require a full blockchain download. This will make the information on this page future-proof.