Suggested clarification on "How Ticket Voting Works"
Being new to the Decred project, I was trying to understand whether an individual DCR holder could only purchase a single ticket for voting, or if it was possible to purchase multiple tickets, representing multiple votes.

Since this is a philosophical issue that is important to the crypcurrency communities in general (i.e. how access to voting procedures is defined, what are the barriers to participation, what are the pros and cons of each system, etc.), I feel that it is important to present to the public an unambiguous description of the voting system. Based on a conversation with the dev community on Riot, I was told that an individual stakeholder can in fact purchase multiple tickets.

I feel that the existing paragraph on this question is not totally clear. As such, I have suggested a new second sentence in the paragraph to make the point absolutely clear.