Fix dark mode only images.
Dark/light only images have been added to mkdocs-material by implementing it in the colour scheme, rather than in the framework itself.
Because we use a custom colour scheme, we need to make a copy of that implementation. Annoying.

This is likely because the framework is not able to auto-detect which colours are light/dark, e.g. given a red scheme and a green scheme, which is which?